+++ src/helpers/SettingsHelper.ts
@@ -77,6 +77,8 @@ export class Settings {
   private static readConfigPromise: Promise<void> | undefined = undefined;
   private static project: Project | undefined = undefined;
   private static configDebouncer = debounceCallback();
+  private static hasExtendedConfig: boolean = false;
+  private static extendedConfig: { extended: any[]; splitted: any[]; dynamic: boolean } = {} as any;
 
   public static async registerCommands() {
     const ext = Extension.getInstance();
@@ -375,11 +377,53 @@ export class Settings {
   }
 
   /**
+   * Safely updates a setting value.
+   *
+   * @param name - The name of the setting.
+   * @param value - The new value for the setting.
+   * @param updateGlobal - Indicates whether to update the global setting or not. Default is `false`.
+   * @returns A promise that resolves when the setting is updated.
+   */
+  public static async safeUpdate<T>(
+    name: string,
+    value: T,
+    updateGlobal: boolean = false
+  ): Promise<void> {
+    if (Settings.hasExtendedConfig) {
+      const configKey = `${CONFIG_KEY}.${name}`;
+
+      if (
+        Settings.extendedConfig.extended.includes(configKey) ||
+        Settings.extendedConfig.splitted.includes(configKey) ||
+        Settings.extendedConfig.dynamic
+      ) {
+        Notifications.warningWithOutput(
+          `Cannot update setting "${configKey}" because you've extended or split the Front Matter CMS configuration. Please manually add your changes. Check the output for the setting update.`
+        );
+
+        Logger.info(`Updating setting: ${configKey}`, 'SETTING');
+        Logger.info(
+          `New value: 
+${JSON.stringify(value, null, 2)}`,
+          'SETTING'
+        );
+        return;
+      }
+    }
+
+    await Settings.update<T>(name, value, updateGlobal);
+  }
+
+  /**
    * String update config setting
    * @param name
    * @param value
    */
-  public static async update<T>(name: string, value: T, updateGlobal: boolean = false) {
+  public static async update<T>(
+    name: string,
+    value: T,
+    updateGlobal: boolean = false
+  ): Promise<void> {
     const fmConfig = await Settings.projectConfigPath();
 
     if (updateGlobal) {
@@ -423,7 +467,7 @@ export class Settings {
 
       const workspaceSettingValue = Settings.hasWorkspaceSettings<ContentType[]>(name);
       if (workspaceSettingValue) {
-        await Settings.update(name, undefined);
+        await Settings.safeUpdate(name, undefined);
       }
 
       // Make sure to reload the whole config + all the data files
@@ -515,7 +559,7 @@ export class Settings {
     customTaxonomies[taxIdx].options.push(option);
     customTaxonomies[taxIdx].options = [...new Set(customTaxonomies[taxIdx].options)];
     customTaxonomies[taxIdx].options = customTaxonomies[taxIdx].options.sort().filter((o) => !!o);
-    await Settings.update(SETTING_TAXONOMY_CUSTOM, customTaxonomies, true);
+    await Settings.safeUpdate(SETTING_TAXONOMY_CUSTOM, customTaxonomies, true);
   }
 
   /**
@@ -532,7 +576,7 @@ export class Settings {
       customTaxonomies[taxIdx].options = options;
     }
 
-    await Settings.update(SETTING_TAXONOMY_CUSTOM, customTaxonomies, true);
+    await Settings.safeUpdate(SETTING_TAXONOMY_CUSTOM, customTaxonomies, true);
   }
 
   /**
@@ -548,8 +592,8 @@ export class Settings {
         const setting = Settings.config.inspect(settingName);
 
         if (setting && typeof setting.workspaceValue !== 'undefined') {
-          await Settings.update(settingName, setting.workspaceValue, true);
-          await Settings.update(settingName, undefined);
+          await Settings.safeUpdate(settingName, setting.workspaceValue, true);
+          await Settings.safeUpdate(settingName, undefined);
         }
       }
     }
@@ -652,6 +696,13 @@ export class Settings {
    */
   private static async readConfig() {
     try {
+      Settings.hasExtendedConfig = false;
+      Settings.extendedConfig = {
+        extended: [],
+        splitted: [],
+        dynamic: false
+      };
+
       const fmConfig = await Settings.projectConfigPath();
       if (fmConfig && (await existsAsync(fmConfig))) {
         const localConfig = await readFileAsync(fmConfig, 'utf8');
@@ -671,6 +722,8 @@ export class Settings {
       );
       if (configFiles.length === 0) {
         Logger.info(`No ".frontmatter/config" config files found.`);
+      } else {
+        Settings.hasExtendedConfig = true;
       }
 
       // Sort the files by fsPath
@@ -707,6 +760,8 @@ export class Settings {
                     );
 
                     if (dynamicConfig) {
+                      Settings.hasExtendedConfig = true;
+                      Settings.extendedConfig.dynamic = true;
                       Settings.globalConfig = dynamicConfig;
                       Logger.info(`Dynamic config file loaded`);
                     }
@@ -738,11 +793,17 @@ export class Settings {
       return;
     }
 
+    Settings.hasExtendedConfig = true;
     const originalConfig = Object.assign({}, Settings.globalConfig);
     const extendsConfig: string[] = Settings.globalConfig[extendsConfigName];
     for (const externalConfig of extendsConfig) {
       if (externalConfig.endsWith(`.json`)) {
         const config = await Settings.getExternalConfig(externalConfig);
+
+        // Store the config keys
+        const keys = Object.keys(config);
+        Settings.extendedConfig.extended.push(...keys);
+
         await Settings.extendConfig(config, originalConfig);
       }
     }
@@ -777,6 +838,7 @@ export class Settings {
         Settings.globalConfig = {};
       }
 
+      Settings.extendedConfig.splitted.push(`${CONFIG_KEY}.${relSettingName}`);
       Settings.updateGlobalConfigSetting(relSettingName, configJson, configFilePath, filePath);
     } catch (e) {
       Logger.error(`Error reading config file: ${configFile.fsPath}`);