+++ src/models/ContentFolder.ts
@@ -10,8 +10,8 @@ export interface ContentFolder {
   filePrefix?: string;
   contentTypes?: string[];
   originalPath?: string;
-  $schema?: string;
-  extended?: boolean;
+  $schema?: string; // Extended config
+  extended?: boolean; // Extended config
 
   locale?: string;
   localeTitle?: string;